moved unending resolve to shared warlock spells, renamed shared/talents to shared/spells
