+++ src/analysis/retail/warlock/shared/Abilities.ts
@@ -34,6 +34,20 @@ class Abilities extends CoreAbilities {
         },
         buffSpellId: TALENTS.DARK_PACT_TALENT.id,
       },
+      {
+        spell: SPELLS.UNENDING_RESOLVE.id,
+        buffSpellId: SPELLS.UNENDING_RESOLVE.id,
+        category: SPELL_CATEGORY.DEFENSIVE,
+        cooldown: 180 - (combatant.hasTalent(TALENTS.DARK_ACCORD_TALENT) ? 45 : 0),
+        gcd: null,
+        castEfficiency: {
+          suggestion: true,
+          importance: ISSUE_IMPORTANCE.MINOR,
+          recommendedEfficiency: 0.33,
+          averageIssueEfficiency: 0.2,
+          majorIssueEfficiency: 0.1,
+        },
+      },
 
       // Utility
       {
