+++ src/analysis/retail/warlock/shared/index.ts
@@ -5,5 +5,6 @@ export { SoulShardIcon };
 export { default as GrimoireOfSacrifice } from './GrimoireOfSacrifice';
 export { default as GrimoireOfSacrificeNormalizer } from './GrimoireOfSacrificeNormalizer';
 export { default as WarlockMissingDotApplyDebuffPrePull } from './WarlockMissingDotApplyDebuffPrePull';
-export { default as DemonicCircle } from './talents/DemonicCircle';
-export { default as DarkPact } from './talents/DarkPact';
+export { default as DemonicCircle } from './spells/DemonicCircle';
+export { default as DarkPact } from './spells/DarkPact';
+export { default as UnendingResolve } from './spells/UnendingResolve';
